debezium/dbz#1668 Move DebeziumServerSink interface to api package
