+++ debezium-server-http/src/main/java/io/debezium/server/http/HttpChangeConsumer.java
@@ -38,7 +38,7 @@ import io.debezium.engine.DebeziumEngine;
 import io.debezium.metadata.ComponentMetadata;
 import io.debezium.metadata.ComponentMetadataFactory;
 import io.debezium.server.BaseChangeConsumer;
-import io.debezium.server.DebeziumServerSink;
+import io.debezium.server.api.DebeziumServerSink;
 import io.debezium.server.http.jwt.JWTAuthenticatorBuilder;
 import io.debezium.server.http.oauth2.OAuth2AuthenticatorBuilder;
 import io.debezium.server.http.webhooks.StandardWebhooksAuthenticatorBuilder;
