+++ debezium-server-instructlab/src/main/java/io/debezium/server/instructlab/InstructLabSinkConsumer.java
@@ -43,7 +43,7 @@ import io.debezium.engine.DebeziumEngine.RecordCommitter;
 import io.debezium.metadata.ComponentMetadata;
 import io.debezium.metadata.ComponentMetadataFactory;
 import io.debezium.server.BaseChangeConsumer;
-import io.debezium.server.DebeziumServerSink;
+import io.debezium.server.api.DebeziumServerSink;
 import io.debezium.util.Strings;
 
 /**
