+++ debezium-server-sqs/src/main/java/io/debezium/server/sqs/SqsChangeConsumer.java
@@ -28,7 +28,7 @@ import io.debezium.engine.DebeziumEngine.RecordCommitter;
 import io.debezium.metadata.ComponentMetadata;
 import io.debezium.metadata.ComponentMetadataFactory;
 import io.debezium.server.BaseChangeConsumer;
-import io.debezium.server.DebeziumServerSink;
+import io.debezium.server.api.DebeziumServerSink;
 import io.debezium.util.Clock;
 import io.debezium.util.Metronome;
